+++ .gitignore
@@ -1,8 +1,9 @@
-.env
 drivers/
 secrets/
 target/
 logs/
 backup/
 venv/
+__pycache__/
 *.java
+.env
